requirements.txt
matplotlib==3.8.0
numpy==1.26.0
pygame==2.5.2
PyYAML==6.0.1
